refactor: Rename `lib/plugins/aws/deploy/lib/extendedValidate.js`
