+++ lib/plugins/aws/deploy/index.js
@@ -2,7 +2,7 @@
 
 const _ = require('lodash');
 const writeServiceOutputs = require('../../../cli/write-service-outputs');
-const extendedValidate = require('./lib/extendedValidate');
+const extendedValidate = require('./lib/extended-validate');
 const setBucketName = require('../lib/set-bucket-name');
 const checkForChanges = require('./lib/check-for-changes');
 const monitorStack = require('../lib/monitor-stack');
