+++ lib/plugins/aws/deploy/lib/extended-validate.js
