+++ test/unit/lib/plugins/aws/deploy/lib/extended-validate.test.js
